+++ examples/lists.py
@@ -9,8 +9,8 @@ Code usage::
 
 Command line usage::
 
-    $ lists.py 2 --numbers 1.2 3.4
-    $ lists.py --numbers 1.2 3.4 -- 2
+    $ python lists.py 2 --numbers 1.2 3.4
+    $ python lists.py --numbers 1.2 3.4 -- 2
 """
 import defopt