+++ docs/changelog.rst
@@ -1,6 +1,11 @@
 Changelog
 =========
 
+development
+-----------
+
+* Added support for type annotations
+
 1.1.0 (2016-02-21)
 ------------------